Replace src in html

+++ .github/scripts/dedup.py
@@ -2,6 +2,8 @@ import pathlib
 import sys
 import os
 
+from bs4 import BeautifulSoup
+
 lines = open(sys.argv[1]).readlines()
 
 duplicates = {}
@@ -12,11 +14,50 @@ for line in lines:
     else:
         duplicates[sha] = [pathlib.Path(path)]
 
+orgMap = {}
 for (first, *rest) in duplicates.values():
     for item in rest:
-        item.unlink()
-        relpath = pathlib.Path(os.path.relpath(first, item.parent))
-        item.symlink_to(relpath)
+        orgMap[item] = first
 
-    if len(rest) > 0:
-        print(f"Created {len(rest)} symlinks to {first}")
+def replaceSource(items):
+    for item in items:
+        src = item["src"]
+        if reportDir / src in orgMap:
+            org = orgMap[reportDir/src]
+            rel = os.path.relpath(org, (reportDir/src).parent)
+            item["src"] = f"/{rel}"
+
+
+root = pathlib.Path(sys.argv[2])
+reports = list(root.glob("**/report*.html"))
+
+for report in reports:
+    reportDir = report.parent
+
+    with open(report) as f:
+        html = f.read()
+    soup = BeautifulSoup(html, 'html.parser')
+
+    replaceSource(soup.find_all('img'))
+    replaceSource(soup.find_all('script'))
+
+    with open(report, 'w') as f:
+        f.write(soup.prettify())
+
+for (first, *rest) in duplicates.values():
+    if first.suffix in [".png", ".js"]:
+        for item in rest:
+            item.unlink()
+
+        if len(rest) > 0:
+            print(f"Removed {len(rest)} duplicates of {first}")
+
+if False:
+    for (first, *rest) in duplicates.values():
+        for item in rest:
+            item.unlink()
+            relpath = pathlib.Path(os.path.relpath(first, item.parent))
+            item.symlink_to(relpath)
+
+        if len(rest) > 0:
+            print(f"Created {len(rest)} symlinks to {first}")
